+++ iti-react/src/__tests__/Validation/UseValidation.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import { CancellablePromise } from '@interface-technologies/iti-react-core'
+import { CancellablePromise } from 'real-cancellable-promise'
 import { ValidatedInput } from '../../Inputs'
 
 it('does not update the component asynchoronously when there is no asyncValidator', async () => {