+++ test-website/src/Components/Layout.tsx
@@ -11,12 +11,7 @@ import {
     ItiReactCoreContext,
 } from '@interface-technologies/iti-react'
 import { LoadingIcon } from 'Components/Icons'
-import ReactHintFactory from 'react-hint'
 import { isConnectionError, store, errorActions } from '_Redux'
-
-const ReactHint = ReactHintFactory(React)
-
-export let forceUpdateTooltips: () => void = () => {}
 
 const itiReactContextData: ItiReactContextData = {
     ...defaultItiReactContextData,
@@ -43,18 +38,6 @@ export function Layout(props: LayoutProps) {
 
     return (
         <div className="layout">
-            <ReactHint
-                events
-                delay={100}
-                attribute="data-tooltip"
-                ref={(ref: any) => {
-                    if (ref) {
-                        forceUpdateTooltips = () => {
-                            ref.forceUpdate()
-                        }
-                    }
-                }}
-            />
             <Navbar activeNavbarLink={activeNavbarLink} />
             <div className="body-container-wrapper">
                 <div className="container">