+++ iti-react-core/Changelog.md
@@ -1,5 +1,6 @@
 ﻿# 3.0.0
 
+- Replace `CancellablePromise` with `real-cancellable-promise` (works mostly the same)
 - Rename query hooks to avoid name collision with `react-query`:  
 	- `useQuery` -> `useSimpleQuery`  
 	- `useParameterlessQuery` -> `useSimpleParameterlessQuery`